requirements_old.txt
attrs==19.3.0
blinker==1.4
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
email-validator==1.1.1
firebase-admin==4.3.0
Flask==1.1.2
Flask-WTF==0.14.3
google-api-core==1.19.1
google-api-python-client==1.9.1
google-auth==1.16.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
google-cloud-secret-manager==1.0.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.29.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
msgpack==1.0.0
packaging==20.4
pluggy==0.13.1
protobuf==3.12.2
py==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==5.4.3
pytest-dotenv==0.5.2
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
